# Faster scanning of memcache responses