@pior 2024-10-09
This looks like a great improvement, I think @bboreham addressed both comments, do you think you can take another look @bradfitz ?